tueresmuyinteligenterafael

+++ analisis/models/grupos_analisis_rel.py
@@ -2,8 +2,8 @@ from analisis import db
 import datetime
 class GruposAnalisisRel(db.Model):
     __tablename__='grupos_analisis_rel'
-    gana_grupo_id_fk =db.Column(db.Integer(10))
-    gana_ana_id_fk =db.Column(db.Integer(10))
+    gana_grupo_id_fk =db.Column(db.Integer)
+    gana_ana_id_fk =db.Column(db.Integer)
 
 
     def __init__(self,gana_grupo_id_fk, gana_ana_id_fk )->None:
